+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A new feature for automated calculation of HOC
 - A new feature for two-stage calculation of interactions
 - A file for all references and citations
+- Descriptive statistics for item and construct data
+- S3 print method for class "table_output" for printing generic tables
 
 ### Changed
 - Fixtures for evaluating bootstrap HTMT for versions of R < 3.6.0